Cleaning up Stat Helpers and writing Stat Helpers Test
